proper validation

+++ app/public/javascript/main.js
@@ -45,13 +45,10 @@ const appendToTable = (request, div) => {
 }
 
 //===========================================================================
-    // variables
+// variables
 
 // bamazon-express.heroku.com/
 const currentURL = window.location.origin;
-
-// last row id
-const lastId = $.get(currentURL + "/fromdatabaselastrow", (request) => { return request[0].id })
 
 //===========================================================================
 // add new item to the database button submit
@@ -74,19 +71,18 @@ $(".new-item").on("submit", (event) => {
 
     // validation
     if (quantity > 200 || quantity <= 0 ||
-         product.length > 255 || Number(price) <= 0 ||
-          Number(price) > 10000 || department === 'Choose...') {
+        product.length > 255 || product.length === 0 || Number(price) <= 0 ||
+        Number(price) > 10000 || department === 'Choose...') {
 
         // redirect to the error page 
         return location.href = currentURL + "/error";
 
     } else {
-        
         // Send the POST request.
         $.ajax("/posttodatabase", {
             type: "POST",
             data: newItem
-        }).then( () => {
+        }).then(() => {
             // redirect to success page
             location.href = currentURL + "/newItemSuccess";
         });
@@ -96,6 +92,9 @@ $(".new-item").on("submit", (event) => {
 //===========================================================================
 // update quantity button submit
 
+// last row id
+const lastId = $.get(currentURL + '/fromdatabaselastrow', (request) => { return request })
+
 $(".update-quantity").on("submit", (event) => {
     // preventDefault on a submit event.
     event.preventDefault();
@@ -104,26 +103,28 @@ $(".update-quantity").on("submit", (event) => {
         id: $("#id").val().trim(),
         quantity: $("#quantity").val().trim()
     };
+
     const id = updateQuantity.id
     const quantity = updateQuantity.quantity
 
     // validation
-    if (id > lastId || id <= 0 || quantity > 200 || quantity <= 0) {
+    if (id > lastId.responseJSON[0].id || id <= 0 || quantity <= 0 || quantity > 200) {
+
+        // redirect to the error page 
         return location.href = currentURL + "/error";
 
     } else {
-
-    // Send the PUT request.
+        // Send the PUT request.
         $.ajax("/addquantity", {
             type: "PUT",
             data: updateQuantity
-        }).then( () => {
+        }).then(() => {
             // redirect to success page
             location.href = currentURL + "/addinventorysuccess";
-            }
-        );
+        });
     }
 })
+
 //===========================================================================
 // delete a item :(
 
@@ -131,24 +132,24 @@ $(".delete-item").on("submit", (event) => {
     // preventDefault on a submit event.
     event.preventDefault();
 
-    const itemId = { id : $("#id").val().trim() } 
+    const itemId = { id: $("#id").val().trim() }
     const ID = itemId.id
 
-    // WONT LET ME VALIDATE?????!!
-    // if ( ID > lastId ) {
-    //     return location.href = currentURL + "/error";
-    // } else {
+    // validation
+    if (ID > lastId.responseJSON[0].id || ID <= 0) {
 
+        return location.href = currentURL + "/error";
+
+    } else {
         // Send the DELETE request.
         $.ajax("/delete", {
             type: "DELETE",
             data: itemId
-        }).then( () => {
+        }).then(() => {
             // redirect to success page
             console.log(ID)
             $("#errorDiv").html("<h3 style = 'color: red;'> OMG <br>you just deleted item number " + ID +
-                                    "!!!!!!!!<h3><h6>how could you..</h6>")
+                "!!!!!!!!<h3><h6>how could you..</h6>")
         });
-    // }
+    }
 })
-
